+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Billboard: Dark Theme & Flatly Theme
 - Billboard: Tick Value [#26](https://github.com/Geuthur/aa-ledger/issues/26)
 - Billboard: Hourly Statistics [#26](https://github.com/Geuthur/aa-ledger/issues/26)
+- Billboard: RattingBar Size Handler
 - add_info_to_context: Theme Check cause NIGHT_MODE won't work with AA v4
 - Create Missing Char Task for alts and mains function
 - Character Ledger: Create Graphical Overview for Single Lookup